// eca-web/src/main/resources/src/app/common/model/experiment-type.enum.ts
export enum ExperimentType {
  NEURAL_NETWORKS = 'NEURAL_NETWORKS',
  HETEROGENEOUS_ENSEMBLE = 'HETEROGENEOUS_ENSEMBLE',
  MODIFIED_HETEROGENEOUS_ENSEMBLE = 'MODIFIED_HETEROGENEOUS_ENSEMBLE',
  ADA_BOOST = 'ADA_BOOST',
  STACKING = 'STACKING',
  KNN = 'KNN',
  RANDOM_FORESTS = 'RANDOM_FORESTS',
  STACKING_CV = 'STACKING_CV',
  DECISION_TREE = 'DECISION_TREE'
}